[AST/libclang] Fix the selector locations that are reported for a method definition that has its '{' attached to the method name without a space.
With a method like:

-(id)meth{
.....
}

the logic in ObjCMethodDecl that determined the selector locations got
confused because it was initialized based on an end location for '{' but
that end location changed to '}' after the method was finished.

Fix this by having an immutable end location for the declarator and
for getLocEnd() get the end location from the body itself.

Fixes rdar://11659739.

git-svn-id: https://llvm.org/svn/llvm-project@158583 91177308-0d34-0410-b5e6-96231b3b80d8