Add support for building the ASan instrumentation unit tests with the just-built Clang binary, and linking them against the just-built ASan runtime.
This is *very* brittle. I expect it will require tweaking, and I've
pro-actively disabled it on non-Unix builds and on cross-builds.

It is also currently missing dependency edges on GoogleTest header files
and a few other corner cases, but those can be fixed. This is the major
milestone of a mini-bootstrap-like build of the unittest.

git-svn-id: https://llvm.org/svn/llvm-project@159255 91177308-0d34-0410-b5e6-96231b3b80d8