[LNT] Added code to View2D.js which allows one to highlight the area in between two revisions in a graph.
This will be used to highlight the area in between two revisions in a run graph.
The idea is that the graph shows you the entire graph (even outside of the revision
range in question), so you can see if a regression recovered. This becomes an
issue when one can not easily visually identify the part of the line graph that
actually corresponds to the revisions you are looking for regressions in.

git-svn-id: https://llvm.org/svn/llvm-project@161212 91177308-0d34-0410-b5e6-96231b3b80d8