Experiment with a gdb/gcc dejagnu test runner in the lab.
This is a bit cheesy just slapping it in as an option to the ClangBuilder, so
if anyone has ideas how they'd like to see the variations, extension points, 
etc, implemented - feel free to let me know in post-commit review.

I've done some experimentation with this locally, but without a mac box to run
it against, so I'll be testing this in the lab with incremental/experimental
commits for now.


git-svn-id: https://llvm.org/svn/llvm-project@161799 91177308-0d34-0410-b5e6-96231b3b80d8