[LNT] Fixed a bug where after choosing a different field to view, using the right click show graph functionality failed to work.
The specific issue occured because in a previous commit I changed the url field encoding to the actual name of the field so that
the code was more general. I failed to change the code which generated the correct url for the graph for a specific test. Additionally,
even on the first load the index generated could be incorrect since I used field.id instead of field.index.

git-svn-id: https://llvm.org/svn/llvm-project@162706 91177308-0d34-0410-b5e6-96231b3b80d8