[LNT] Updated v4db.py so that an engine is reused in between requests.
The reason to do this is that currently for every request we create a new engine,
which interferes with SQLAlchemy's internal connection caching. Specifically,
every time we create a new v4db, a new connection pool is created which
will not recycle its connections implying a continually increasing
amount of connections which never close, overwhelming the database.

git-svn-id: https://llvm.org/svn/llvm-project@163169 91177308-0d34-0410-b5e6-96231b3b80d8