Take another crack at stabilizing the emission order of analyzer diagnostics without using FoldingSetNodeIDs.  This is done by doing a complete recursive comparison of the PathDiagnostics.
Note that the previous method of comparing FoldingSetNodeIDs did
not end up relying on unstable things such as pointer addresses, so
I suspect this may still have some issues on various buildbots because
I'm not sure if the true source of non-determinism has been eliminated.
The tests pass for me, so the only way to know is to commit this change
and see what happens.

git-svn-id: https://llvm.org/svn/llvm-project@163489 91177308-0d34-0410-b5e6-96231b3b80d8