[LNT] Added initial versions of server/reporting/runs.py templates for both html and txt output formats.
Additionally:

1. Updated setup.py to ensure that said templates are copied
into any build dist/eggs.

2. Added filter urlencode to lnt.server.ui.filters. This is just a
filter wrapper around urllib.urlencode.

3. Added filter format_or_default to lnt.server.ui.filters. This is just
a formatting filter which takes in 3 arguments, a format, a tuple of
args, and a default value. If the tuple is not None, then the filter
returns format % tuple. If tuple is None, the filter returns default. I
used this to refactor out formatting like format_or_default('%s - %s',
args, '-') so that a '-' will be used if args is None.

git-svn-id: https://llvm.org/svn/llvm-project@163779 91177308-0d34-0410-b5e6-96231b3b80d8