Clear UNIVERSAL_SDK_PATH setting when building host tools.  <rdar://12360497>
I also moved the SDKROOT setting into the make flags, since clearing it from
the environment isn't good enough to override a setting on the make command
line.  That hasn't been a problem but it could be, and it's good to be
consistent with the way UNIVERSAL_SDK_PATH is handled.

git-svn-id: https://llvm.org/svn/llvm-project@164565 91177308-0d34-0410-b5e6-96231b3b80d8