Suppress three signal related gdb tests since they don't tell us anything about Clang.
And these are failing on the public buildbot. In theory we could just delete
all the tests that aren't actually about debug information, but that's more
work than anyone's willing to put into it at the moment (& wouldn't provide
much value other than a minor speed up to the gdb bots).


git-svn-id: https://llvm.org/svn/llvm-project@167169 91177308-0d34-0410-b5e6-96231b3b80d8