<rdar://problem/12645617> <rdar://problem/12153915>  (partial fix)
Remove an assert and place an error message instead so we don't crash when we run into a type tag that we don't recognize. We will now emit a warning so that hopefully we can get a bug report that has example code that shows what we are missing.

Also fixed a case when trying to unique one type to another where we would confuse concrete instances of methods with their definitions and end up not correctly registering the types.



git-svn-id: https://llvm.org/svn/llvm-project@167557 91177308-0d34-0410-b5e6-96231b3b80d8